+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # CHANGELOG
 
+## 1.10.0
+
+* perf: warmable cache
+
 ## 1.9.4
 
 * deps: fix `symfony/stopwatch` dependency
 * fix: `Map` now works with unpromoted constructor parameters
 * fix: restore throwing `NewInstanceReturnedButCannotBeSetOnTargetException`
 * fix: do not run DebugPass if symfony/stopwatch is not present
-* deps: `symfony/stopwatch` is required by non-dev
-* perf: warmable cache
 
 ## 1.9.3